pyproject.toml
[build-system]
requires = [
    "setuptools",
    "versioneer[toml]==0.29",
    # tomli is used by versioneer
    "tomli; python_version < '3.11'",
    # ninja needs to be provided by the environment
    "cmake",
]
build-backend = "setuptools.build_meta"

[project]
name = "pymgl"
dynamic = ["version"]
authors = [{ name = "Brendan C. Ward", email = "bcward@astutespruce.com" }]
license = { file = "LICENSE" }
description = "Python wrapper for MapLibre Native"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: GIS",
]
requires-python = ">=3.9"
dependencies = ["packaging"]

[project.optional-dependencies]
dev = ["numpy", "Pillow", "pre-commit", "ruff", "setuptools", "versioneer"]
test = [
    "numpy",
    "Pillow",
    "pixelmatch",
    "python-dotenv",
    "pytest",
    "pytest-cov",
]
bench = [
    "numpy",
    "Pillow",
    "pixelmatch",
    "python-dotenv",
    "pytest",
    "pytest-benchmark",
    "requests",
]
slownetwork = ["poorconn"]

[project.urls]
Home = "https://github.com/brendan-ward/pymgl"
Repository = "https://github.com/brendan-ward/pymgl"

[tool.setuptools]
packages = ["pymgl", "pymgl.tests", "pymgl.tests.fixtures"]

[tool.uv]
package = false

[tool.versioneer]
VCS = "git"
style = "pep440"
versionfile_source = "pymgl/_version.py"
versionfile_build = "pymgl/_version.py"
parentdir_prefix = "pymgl-"
tag_prefix = "v"

[tool.pytest.ini_options]
testpaths = ["pymgl/tests"]
